Create Simple Linear Regression
This program was able to build a simple linear regression model that describes the training set and also predict future data (such as the test set). By looking at the Test plot, one can see that the Simple Linear Regression did a good job predicting not only the training data but also the test data as the points lie  close to the prediction line.